Add files via upload
This is a headless browser tester that I wrote for a website at work. It clicks every link on the homepage, and it checks for the error page that would be displayed on the site it was intended for.

This could easily be manipulated to test other sites. With some changes, it could also have utility in something like hunting bug bounties.